rename modules
